requirements.txt
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
PyGithub>=1.59.0
